nvme: move get-log page commands into a plugin
There are too many top-level commands cluttering the space. Move the
specialized get-log page commands into a plugin, e.g.,

  nvme smart-log -> nvme log smart

The main command 'nvme get-log' remains available at the top level.
