+++ src/nvme-builtin.h
@@ -34,40 +34,46 @@ COMMAND_LIST(
 	ENTRY("detach-ns", "Detaches a namespace from requested controller(s)", detach_ns)
 	ENTRY("get-ns-id", "Retrieve the namespace ID of opened block device", get_ns_id)
 	ENTRY("get-log", "Generic NVMe get log, returns log in raw format", get_log)
-	ENTRY("telemetry-log", "Retrieve FW Telemetry log write to file", get_telemetry_log)
-	ENTRY("fw-log", "Retrieve FW Log, show it", get_fw_log)
-	ENTRY("changed-ns-list-log", "Retrieve Changed Attached Namespace List, show it", get_changed_attach_ns_list_log)
-	ENTRY("smart-log", "Retrieve SMART Log, show it", get_smart_log)
-	ENTRY("ana-log", "Retrieve ANA Log, show it", get_ana_log)
-	ENTRY("error-log", "Retrieve Error Log, show it", get_error_log)
-	ENTRY("effects-log", "Retrieve Command Effects Log, show it", get_effects_log)
-	ENTRY("endurance-log", "Retrieve Endurance Group Log, show it", get_endurance_log)
-	ENTRY("predictable-lat-log", "Retrieve Predictable Latency per Nvmset Log, show it", get_pred_lat_per_nvmset_log)
-	ENTRY("pred-lat-event-agg-log", "Retrieve Predictable Latency Event Aggregate Log, show it", get_pred_lat_event_agg_log)
-	ENTRY("persistent-event-log", "Retrieve Persistent Event Log, show it", get_persistent_event_log)
-	ENTRY("endurance-event-agg-log", "Retrieve Endurance Group Event Aggregate Log, show it", get_endurance_event_agg_log)
-	ENTRY("lba-status-log", "Retrieve LBA Status Information Log, show it", get_lba_status_log)
-	ENTRY("resv-notif-log", "Retrieve Reservation Notification Log, show it", get_resv_notif_log)
-	ENTRY("boot-part-log", "Retrieve Boot Partition Log, show it", get_boot_part_log)
-	ENTRY("phy-rx-eom-log", "Retrieve Physical Interface Receiver Eye Opening Measurement, show it", get_phy_rx_eom_log)
+#ifdef CONFIG_DEPRECATED_CMDS
+	ENTRY("telemetry-log", "Retrieve FW Telemetry log write to file (deprecated, use 'nvme log telemetry')", get_telemetry_log)
+	ENTRY("fw-log", "Retrieve FW Log, show it (deprecated, use 'nvme log fw')", get_fw_log)
+	ENTRY("changed-ns-list-log", "Retrieve Changed Attached Namespace List, show it (deprecated, use 'nvme log changed-ns-list')", get_changed_attach_ns_list_log)
+	ENTRY("smart-log", "Retrieve SMART Log, show it (deprecated, use 'nvme log smart')", get_smart_log)
+	ENTRY("ana-log", "Retrieve ANA Log, show it (deprecated, use 'nvme log ana')", get_ana_log)
+	ENTRY("error-log", "Retrieve Error Log, show it (deprecated, use 'nvme log error')", get_error_log)
+	ENTRY("effects-log", "Retrieve Command Effects Log, show it (deprecated, use 'nvme log effects')", get_effects_log)
+	ENTRY("endurance-log", "Retrieve Endurance Group Log, show it (deprecated, use 'nvme log endurance')", get_endurance_log)
+	ENTRY("predictable-lat-log", "Retrieve Predictable Latency per Nvmset Log, show it (deprecated, use 'nvme log predictable-lat')", get_pred_lat_per_nvmset_log)
+	ENTRY("pred-lat-event-agg-log", "Retrieve Predictable Latency Event Aggregate Log, show it (deprecated, use 'nvme log pred-lat-event-agg')", get_pred_lat_event_agg_log)
+	ENTRY("persistent-event-log", "Retrieve Persistent Event Log, show it (deprecated, use 'nvme log persistent-event')", get_persistent_event_log)
+	ENTRY("endurance-event-agg-log", "Retrieve Endurance Group Event Aggregate Log, show it (deprecated, use 'nvme log endurance-event-agg')", get_endurance_event_agg_log)
+	ENTRY("lba-status-log", "Retrieve LBA Status Information Log, show it (deprecated, use 'nvme log lba-status')", get_lba_status_log)
+	ENTRY("resv-notif-log", "Retrieve Reservation Notification Log, show it (deprecated, use 'nvme log resv-notif')", get_resv_notif_log)
+	ENTRY("boot-part-log", "Retrieve Boot Partition Log, show it (deprecated, use 'nvme log boot-part')", get_boot_part_log)
+	ENTRY("phy-rx-eom-log", "Retrieve Physical Interface Receiver Eye Opening Measurement, show it (deprecated, use 'nvme log phy-rx-eom')", get_phy_rx_eom_log)
+#endif /* CONFIG_DEPRECATED_CMDS */
 	ENTRY("get-feature", "Get feature and show the resulting value", get_feature)
 	ENTRY("device-self-test", "Perform the necessary tests to observe the performance", device_self_test)
-	ENTRY("self-test-log", "Retrieve the SELF-TEST Log, show it", self_test_log)
+#ifdef CONFIG_DEPRECATED_CMDS
+	ENTRY("self-test-log", "Retrieve the SELF-TEST Log, show it (deprecated, use 'nvme log self-test')", self_test_log)
+#endif /* CONFIG_DEPRECATED_CMDS */
 	ENTRY("supported-log-pages", "Retrieve the Supported Log pages details, show it", get_supported_log_pages)
-	ENTRY("fid-support-effects-log", "Retrieve FID Support and Effects log and show it", get_fid_support_effects_log)
-	ENTRY("mi-cmd-support-effects-log", "Retrieve MI Command Support and Effects log and show it", get_mi_cmd_support_effects_log)
-	ENTRY("media-unit-stat-log", "Retrieve the configuration and wear of media units, show it", get_media_unit_stat_log)
-	ENTRY("supported-cap-config-log", "Retrieve the list of Supported Capacity Configuration Descriptors", get_supp_cap_config_log)
-	ENTRY("mgmt-addr-list-log", "Retrieve Management Address List Log, show it", get_mgmt_addr_list_log)
-	ENTRY("rotational-media-info-log", "Retrieve Rotational Media Information Log, show it", get_rotational_media_info_log)
-	ENTRY("changed-alloc-ns-list-log", "Retrieve Changed Allocated Namespace List, show it", get_changed_alloc_ns_list_log)
-	ENTRY("dispersed-ns-participating-nss-log", "Retrieve Dispersed Namespace Participating NVM Subsystems Log, show it", get_dispersed_ns_participating_nss_log)
-	ENTRY("reachability-groups-log", "Retrieve Reachability Groups Log, show it", get_reachability_groups_log)
-	ENTRY("reachability-associations-log", "Retrieve Reachability Associations Log, show it", get_reachability_associations_log)
-	ENTRY("host-discovery-log", "Retrieve Host Discovery Log, show it", get_host_discovery_log)
-	ENTRY("ave-discovery-log", "Retrieve AVE Discovery Log, show it", get_ave_discovery_log)
-	ENTRY("pull-model-ddc-req-log", "Retrieve Pull Model DDC Request Log, show it", get_pull_model_ddc_req_log)
-	ENTRY("power-measurement-log", "Retrieve Power Measurement Log, show it", get_power_measurement_log)
+#ifdef CONFIG_DEPRECATED_CMDS
+	ENTRY("fid-support-effects-log", "Retrieve FID Support and Effects log and show it (deprecated, use 'nvme log fid-support-effects')", get_fid_support_effects_log)
+	ENTRY("mi-cmd-support-effects-log", "Retrieve MI Command Support and Effects log and show it (deprecated, use 'nvme log mi-cmd-support-effects')", get_mi_cmd_support_effects_log)
+	ENTRY("media-unit-stat-log", "Retrieve the configuration and wear of media units, show it (deprecated, use 'nvme log media-unit-stat')", get_media_unit_stat_log)
+	ENTRY("supported-cap-config-log", "Retrieve the list of Supported Capacity Configuration Descriptors (deprecated, use 'nvme log supported-cap-config')", get_supp_cap_config_log)
+	ENTRY("mgmt-addr-list-log", "Retrieve Management Address List Log, show it (deprecated, use 'nvme log mgmt-addr-list')", get_mgmt_addr_list_log)
+	ENTRY("rotational-media-info-log", "Retrieve Rotational Media Information Log, show it (deprecated, use 'nvme log rotational-media-info')", get_rotational_media_info_log)
+	ENTRY("changed-alloc-ns-list-log", "Retrieve Changed Allocated Namespace List, show it (deprecated, use 'nvme log changed-alloc-ns-list')", get_changed_alloc_ns_list_log)
+	ENTRY("dispersed-ns-participating-nss-log", "Retrieve Dispersed Namespace Participating NVM Subsystems Log, show it (deprecated, use 'nvme log dispersed-ns-participating-nss')", get_dispersed_ns_participating_nss_log)
+	ENTRY("reachability-groups-log", "Retrieve Reachability Groups Log, show it (deprecated, use 'nvme log reachability-groups')", get_reachability_groups_log)
+	ENTRY("reachability-associations-log", "Retrieve Reachability Associations Log, show it (deprecated, use 'nvme log reachability-associations')", get_reachability_associations_log)
+	ENTRY("host-discovery-log", "Retrieve Host Discovery Log, show it (deprecated, use 'nvme log host-discovery')", get_host_discovery_log)
+	ENTRY("ave-discovery-log", "Retrieve AVE Discovery Log, show it (deprecated, use 'nvme log ave-discovery')", get_ave_discovery_log)
+	ENTRY("pull-model-ddc-req-log", "Retrieve Pull Model DDC Request Log, show it (deprecated, use 'nvme log pull-model-ddc-req')", get_pull_model_ddc_req_log)
+	ENTRY("power-measurement-log", "Retrieve Power Measurement Log, show it (deprecated, use 'nvme log power-measurement')", get_power_measurement_log)
+#endif /* CONFIG_DEPRECATED_CMDS */
 	ENTRY("set-feature", "Set a feature and show the resulting value", set_feature)
 	ENTRY("set-property", "Set a property and show the resulting value", set_property)
 	ENTRY("get-property", "Get a property and show the resulting value", get_property)
@@ -94,7 +100,9 @@ COMMAND_LIST(
 	ENTRY("write-uncor", "Submit a write uncorrectable command, return results", write_uncor)
 	ENTRY("verify", "Submit a verify command, return results", verify_cmd)
 	ENTRY("sanitize", "Submit a sanitize command", sanitize_cmd)
-	ENTRY("sanitize-log", "Retrieve sanitize log, show it", sanitize_log)
+#ifdef CONFIG_DEPRECATED_CMDS
+	ENTRY("sanitize-log", "Retrieve sanitize log, show it (deprecated, use 'nvme log sanitize')", sanitize_log)
+#endif /* CONFIG_DEPRECATED_CMDS */
 	ENTRY("sanitize-ns", "Submit a sanitize namespace command",
 	      sanitize_ns_cmd)
 	ENTRY("reset", "Resets the controller", reset)
