+++ tests/e2e/nvme_lba_status_log_test.py
@@ -37,13 +37,13 @@ class TestNVMeLbaStatLogCmd(TestNVMe):
         super().tearDown()
 
     def get_lba_stat_log(self):
-        """ Wrapper for executing nvme lba-status-log.
+        """ Wrapper for executing nvme log lba-status.
             - Args:
                 - None
             - Returns:
                 - 0 on success, error code on failure.
         """
-        lba_stat_log_cmd = f"{self.nvme_bin} lba-status-log {self.ctrl}"
+        lba_stat_log_cmd = f"{self.nvme_bin} log lba-status {self.ctrl}"
         return self.exec_cmd(lba_stat_log_cmd)
 
     def test_lba_stat_log(self):
